added a bttn
backspace bttn

+++ CalculatorGUI/src/CalculatorPCK/CalMethodsClass.java
@@ -9,12 +9,12 @@ package CalculatorPCK;
  * @author User
  */
 public class CalMethodsClass {
-    private int num1;
-    private int num2;
-    private int ans;
+    private double num1;
+    private double num2;
+    private double ans;
     private char symbol;
 
-    public CalMethodsClass(int num1, int num2, char symbol) {
+    public CalMethodsClass(double num1, double num2, char symbol) {
         this.num1 = num1;
         this.num2 = num2;
         this.symbol = symbol;
@@ -22,23 +22,23 @@ public class CalMethodsClass {
 
     
 
-    public int getNum1() {
+    public double getNum1() {
         return num1;
     }
 
-    public void setNum1(int num1) {
+    public void setNum1(double num1) {
         this.num1 = num1;
     }
 
-    public int getNum2() {
+    public double getNum2() {
         return num2;
     }
 
-    public void setNum2(int num2) {
+    public void setNum2(double num2) {
         this.num2 = num2;
     }
 
-    public void setAns(int ans) {
+    public void setAns(double ans) {
         this.ans = ans;
     }
 
@@ -50,7 +50,7 @@ public class CalMethodsClass {
         this.symbol = symbol;
     }
 
-    public int getAns() {
+    public double getAns() {
         if(symbol == '+'){
             ans = num1 + num2;
         }
